+++ modules/server/src/main/ts/bedrock/compiler/Webpack.ts
@@ -20,23 +20,7 @@ interface CompileInfo {
   readonly config: webpack.Configuration;
 }
 
-const moduleAvailable = (name: string) => {
-  try {
-    require.resolve(name);
-    return true;
-  } catch (e) {
-    return false;
-  }
-};
-
-const webpackRemap: Array<Record<string, any>> = moduleAvailable('@ephox/swag') ? [
-  {
-    test: /\.(js|mjs|tsx?)$/,
-    use: [ '@ephox/swag/webpack/remapper' ]
-  }
-] : [];
-
-const webpackSharedRules = webpackRemap.concat([
+const webpackSharedRules = ([] as any[]).concat([
   {
     test: /\.js$/,
     resolve: {